+++ test/visualTest.ts
@@ -432,6 +432,112 @@ describe("AsterPlot", () => {
             });
         });
 
+        function timeout(ms: number) {
+            return new Promise(resolve => setTimeout(resolve, ms));
+        }
+
+        const DefaultWaitForRender: number = 500;
+
+        describe("Keyboard Navigation check", () => {
+            it("links should have attributes tabindex=0, role=option and aria-selected=false", (done) => {
+                visualBuilder.updateRenderTimeout(dataView, () => {
+                    visualBuilder.updateFlushAllD3Transitions(dataView);
+                    let nodes = [...visualBuilder.slices];
+                    nodes.forEach((el: Element) => {
+                        expect(el.getAttribute("role")).toBe("option");
+                        expect(el.getAttribute("tabindex")).toBe("0");
+                        expect(el.getAttribute("aria-selected")).toBe("false");
+                    });
+                    done();
+                },);
+            });
+
+            it("enter toggles the correct slice", (done: DoneFn) => {
+                const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
+                visualBuilder.updateRenderTimeout(
+                    dataView,
+                    async () => {
+                        visualBuilder.slices[0].dispatchEvent(enterEvent);
+                        await timeout(DefaultWaitForRender);
+                        expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                        for (const slice of [...visualBuilder.slices]) {
+                            if (slice !== visualBuilder.slices[0]) {
+                                expect(slice.getAttribute("aria-selected")).toBe("false");
+                            }
+                        }
+
+                        visualBuilder.slices[0].dispatchEvent(enterEvent);
+                        await timeout(DefaultWaitForRender);
+                        for (const slice of [...visualBuilder.slices]) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                        }
+
+                        done();
+                    },
+                    2,
+                );
+            });
+        });
+
+        it("space toggles the correct slice", (done: DoneFn) => {
+            const spaceEvent = new KeyboardEvent("keydown", { code: "Space", bubbles: true });
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    visualBuilder.slices[0].dispatchEvent(spaceEvent);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                    for (const slice of [...visualBuilder.slices]) {
+                        if (slice !== visualBuilder.slices[0]) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                        }
+                    }
+
+                    visualBuilder.slices[0].dispatchEvent(spaceEvent);
+                    await timeout(DefaultWaitForRender);
+                    for (const slice of [...visualBuilder.slices]) {
+                        expect(slice.getAttribute("aria-selected")).toBe("false");
+                    }
+
+                    done();
+                },
+                2,
+            );
+        });
+
+        it("tab between slices works", (done: DoneFn) => {
+            const tabEvent = new KeyboardEvent("keydown", { code: "Tab", bubbles: true });
+            const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    visualBuilder.slices[0].dispatchEvent(enterEvent);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                    for (const slice of [...visualBuilder.slices]) {
+                        if (slice !== visualBuilder.slices[0]) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                        }
+                    }
+
+                    visualBuilder.slices[1].dispatchEvent(tabEvent);
+                    await timeout(DefaultWaitForRender);
+
+                    visualBuilder.slices[1].dispatchEvent(enterEvent);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.slices[1].getAttribute("aria-selected")).toBe("true");
+                    for (const slice of [...visualBuilder.slices]) {
+                        if (slice !== visualBuilder.slices[1]) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                        }
+                    }
+
+                    done();
+                },
+                2,
+            );
+        });
+
         describe("Default Legend", () => {
             const defaultLegendLabelFontSize: number = 8;